fix database and json

+++ database/migrations/2016_12_13_043531_create_rates_table.php
@@ -17,7 +17,7 @@ class CreateRatesTable extends Migration
             $table->increments('id');
             $table->integer('book_id')->unsigned();
             $table->integer('user_id')->unsigned();
-            $table->float('point');
+            $table->integer('point');
             $table->timestamps();
         });
     }
